Move Activator, Deactivator & Uninstaller to app directory
