+++ plugin-name/app/class-activator.php
@@ -1,0 +1,37 @@
+<?php
+namespace Plugin_Name\App;
+use \Plugin_Name\App\Models\Activator as Activator_Model;
+/**
+ * Fired during plugin activation
+ *
+ * @link       http://example.com
+ * @since      1.0.0
+ *
+ * @package    Plugin_Name
+ * @subpackage Plugin_Name/App
+ */
+
+/**
+ * Fired during plugin activation.
+ *
+ * This class defines all code necessary to run during the plugin's activation.
+ *
+ * @since      1.0.0
+ * @package    Plugin_Name
+ * @subpackage Plugin_Name/App
+ * @author     Your Name <email@example.com>
+ */
+class Activator {
+
+	/**
+	 * Short Description. (use period)
+	 *
+	 * Long Description.
+	 *
+	 * @since    1.0.0
+	 */
+	public function activate() {
+
+	}
+
+}
