+++ plugin-name/app/class-deactivator.php
@@ -1,27 +1,27 @@
 <?php
-namespace Plugin_Name\Includes;
+namespace Plugin_Name\App;
 
 /**
- * Fired during plugin activation
+ * Fired during plugin deactivation
  *
  * @link       http://example.com
  * @since      1.0.0
  *
  * @package    Plugin_Name
- * @subpackage Plugin_Name/includes
+ * @subpackage Plugin_Name/App
  */
 
 /**
- * Fired during plugin activation.
+ * Fired during plugin deactivation.
  *
- * This class defines all code necessary to run during the plugin's activation.
+ * This class defines all code necessary to run during the plugin's deactivation.
  *
  * @since      1.0.0
  * @package    Plugin_Name
- * @subpackage Plugin_Name/includes
+ * @subpackage Plugin_Name/App
  * @author     Your Name <email@example.com>
  */
-class Activator {
+class Deactivator {
 
 	/**
 	 * Short Description. (use period)
@@ -30,7 +30,7 @@ class Activator {
 	 *
 	 * @since    1.0.0
 	 */
-	public static function activate() {
+	public function deactivate() {
 	}
 
 }
